requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
openai==0.28.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.1.0
yarl==1.9.4
